Oh man, so, I introduced the stats into the app.
After going through some issues I think I might have solved most of it.

Had change the data types in the Stats class from int to String, even though they are all numbers, I figured that was what caused the app to crash.

Renamed the buttons in the Main Activity as they had the other's name and not the one they were supposed to have.